[SKP-147] refac : useAnalyze 훅으로 분리 2
